feat : 유저 권한 수정

+++ src/main/java/ssafy/horong/domain/community/service/CommunityServiceImpl.java
@@ -11,6 +11,8 @@ import org.springframework.transaction.annotation.Transactional;
 import ssafy.horong.api.community.response.GetCommentResponse;
 import ssafy.horong.api.community.response.GetMessageListResponse;
 import ssafy.horong.api.community.response.GetPostResponse;
+import ssafy.horong.common.exception.Board.NotAdminExeption;
+import ssafy.horong.common.exception.Board.NotAuthenticatedException;
 import ssafy.horong.common.util.S3Util;
 import ssafy.horong.common.util.SecurityUtil;
 import ssafy.horong.domain.community.command.*;
@@ -21,6 +23,7 @@ import ssafy.horong.domain.community.entity.Comment;
 import ssafy.horong.domain.community.repository.BoardRepository;
 import ssafy.horong.domain.community.repository.CommentRepository;
 import ssafy.horong.domain.community.repository.MessageRepository;
+import ssafy.horong.domain.member.common.MemberRole;
 import ssafy.horong.domain.member.entity.User;
 import ssafy.horong.domain.member.repository.UserRepository;
 
@@ -43,32 +46,48 @@ public class CommunityServiceImpl implements CommunityService {
     @Override
     @Transactional
     public void createPost(CreatePostCommand command) {
+        if (command.boardType() == BoardType.NOTICE && SecurityUtil.getLoginMemberRole().orElse(MemberRole.USER) != MemberRole.ADMIN) {
+            throw new NotAdminExeption();
+        }
 
         Post post = Post.builder()
                 .author(getCurrentUser())
                 .title(command.title())
                 .content(command.content())
-                .type(BoardType.valueOf(command.boardType().toUpperCase()))
+                .type(command.boardType())
                 .build();
+
         post.updateImages(s3Util.uploardBoardImageToS3(command.images(), post.getId()));
+
         boardRepository.save(post);
     }
 
     @Override
     @Transactional
     public void updatePost(UpdatePostCommand command) {
-        Post existingBoard = boardRepository.findById(command.postId())
+        Post post = boardRepository.findById(command.postId())
                 .orElseThrow(() -> new RuntimeException("게시글이 존재하지 않습니다."));
 
-        existingBoard.setTitle(command.title());
-        existingBoard.setContent(command.title());
+        if (post.getAuthor() == getCurrentUser() || SecurityUtil.getLoginMemberRole().orElse(MemberRole.USER) == MemberRole.ADMIN) {
+            throw new NotAuthenticatedException();
+        }
 
-        boardRepository.save(existingBoard);
+        post.setTitle(command.title());
+        post.setContent(command.content());
+
+        boardRepository.save(post);
     }
 
     @Override
     @Transactional
     public void deletePost(Long id) {
+        Post post = boardRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("게시글이 존재하지 않습니다."));
+
+        if (post.getAuthor() == getCurrentUser() || SecurityUtil.getLoginMemberRole().orElse(MemberRole.USER) == MemberRole.ADMIN) {
+            throw new NotAuthenticatedException();
+        }
+
         log.info("게시글 삭제: {}", id);
         boardRepository.deleteById(id);
     }
@@ -156,6 +175,14 @@ public class CommunityServiceImpl implements CommunityService {
     @Override
     @Transactional
     public void deleteComment(Long commentId) {
+
+        Comment comment = commentRepository.findById(commentId)
+                .orElseThrow(() -> new ResourceNotFoundException("댓글이 존재하지 않습니다."));
+
+        if (comment.getAuthor() == getCurrentUser() || SecurityUtil.getLoginMemberRole().orElse(MemberRole.USER) == MemberRole.ADMIN) {
+            throw new NotAuthenticatedException();
+        }
+
         log.info("댓글 삭제: {}", commentId);
         commentRepository.deleteById(commentId);
     }
@@ -165,6 +192,10 @@ public class CommunityServiceImpl implements CommunityService {
     public void updateComment(UpdateCommentCommand command) {
         Comment existingComment = commentRepository.findById(command.commentId())
                 .orElseThrow(() -> new RuntimeException("댓글이 존재하지 않습니다."));
+
+        if (existingComment.getAuthor() == getCurrentUser() || SecurityUtil.getLoginMemberRole().orElse(MemberRole.USER) == MemberRole.ADMIN) {
+            throw new NotAuthenticatedException();
+        }
 
         existingComment.setContent(command.content());
         commentRepository.save(existingComment);
